added EventQue class
this stores all of the game events and has a public facing API that allows you to call the handles based on a given name.
it currently runs though all of the current events and does them if the name is matching.
adding multiple events of the same name is yields undefined behavior, as I can not guarantee that it will not return after the first call.  [also this is not something that