added a TID (type id) field to the EventHandler struct
[v-a0.1.0]
I also added the TID to the constructor. The TID should always be the SDL_EventType that the event should fire on if it is not the
Event will not fire because the event loop will call based off of this value.
added a callEventByTID this is the same as callEventByName exept that it comares the tid instead of the name.
marked the callEventByName as deprecated.