[v - a 8.22.0] created Counter
this is a counter that you can set the length of and the max length of. when the internal count is less than the length it will trigger a TimerActiveTrigger with the passed in pid as the counter's uid, when the counter is reset (when count is == m_max) it triggers a TimerTriggerReset event with the pid set to the counters uid.
in the example this is used to draw some triangles to the screen when the timer is active an a circle when the timer resets.